requirements.txt
annotated-types==0.7.0
anyio==4.8.0
binary==1.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
lxml==5.3.0
oauthlib==3.2.2
openai==1.61.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
sniffio==1.3.1
snscrape==0.7.0.20230622
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uuid==1.30
uvicorn==0.34.0
python-multipart
cloudinary
supervisor
pydantic>=2.0.0
celery==5.4.0
kombu